refactor(tests): correct namespaces
