+++ PokeApiNet.Tests/IntegrationTests.cs
@@ -1,5 +1,4 @@
-﻿using PokeApiNet.Data;
-using PokeApiNet.Models;
+﻿using PokeApiNet.Models;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
